Image Helper Updates For Cropping
When third argument on sizing images is set to false, it doesn't crop, it resizes down to the largest size possible within the limits and also keeping the ratio. Then a background of white is added and the canvas is resized to match the width and height of the original request.

Signed-off-by: David Thorpe <davzie@davzie.com>